strategies/copy_trade_strategy.rs
use crate::models::copy_trade_target::CopyTradeTarget;
use crate::models::market_conditions::MarketConditions;
use crate::strategies::strategy::Strategy;
use crate::utils::solana::{analyze_transaction, calculate_profit};
use async_trait::async_trait;
use solana_sdk::pubkey::Pubkey;
use std::collections::HashMap;

pub struct CopyTradeStrategy {
    pub rpc_client: solana_client::rpc_client::RpcClient,
    pub trade_threshold: f64,
    pub max_trade_amount: f64,
}

impl CopyTradeStrategy {
    pub fn new(rpc_client: solana_client::rpc_client::RpcClient) -> Self {
        CopyTradeStrategy {
            rpc_client,
            trade_threshold: 0.0,
            max_trade_amount: 0.0,
        }
    }

    pub fn set_trade_threshold(&mut self, trade_threshold: f64) {
        self.trade_threshold = trade_threshold;
    }

    pub fn set_max_trade_amount(&mut self, max_trade_amount: f64) {
        self.max_trade_amount = max_trade_amount;
    }
}

#[async_trait]
impl Strategy for CopyTradeStrategy {
    fn update(&mut self, market_conditions: &MarketConditions) {
        self.trade_threshold = market_conditions.copy_trade_threshold;
        self.max_trade_amount = market_conditions.max_trade_amount;
    }

    async fn find_opportunities(&self, target_accounts: &HashMap<Pubkey, crate::AccountInfo>) -> Vec<CopyTradeTarget> {
        let mut targets = Vec::new();
        
        for (trader_account, account_info) in target_accounts {
            let recent_trades = self.get_recent_trades(trader_account).await;
            
            for trade in recent_trades {
                let trade_amount = trade.token_amount;
                let target_token = trade.token_mint;
                
                if trade_amount >= self.trade_threshold && trade_amount <= self.max_trade_amount {
                    let copy_trade_target = CopyTradeTarget {
                        trader_account: *trader_account,
                        target_token,
                        trade_amount,
                    };
                    targets.push(copy_trade_target);
                }
            }
        }

        targets
    }
}

impl CopyTradeStrategy {
    async fn get_recent_trades(&self, trader_account: &Pubkey) -> Vec<crate::models::trade::Trade> {
        let trades = Vec::new();
        trades
    }
    
    async fn execute_copy_trade(&self, target: &CopyTradeTarget) -> Option<String> {
        let transaction = self.build_copy_trade_transaction(target);
        let signature = self.rpc_client.send_transaction(&transaction).await;
        
        if let Ok(signature) = signature {
            let transaction_info = self.rpc_client.get_transaction(&signature).await.unwrap();
            let profit = calculate_profit(&transaction_info).unwrap();
            println!("Copy trade transaction executed. Profit: {}", profit);
            Some(signature.to_string())
        } else {
            println!("Failed to execute copy trade transaction");
            None
        }
    }
    
    fn build_copy_trade_transaction(&self, target: &CopyTradeTarget) -> solana_sdk::transaction::Transaction {
        let transaction = solana_sdk::transaction::Transaction::new_with_payer(
            &[],
            Some(&self.rpc_client.payer_pubkey()),
        );
        transaction
    }
}